services/binary-analysis/requirements-simple.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
aiofiles==23.2.1
pydantic==2.5.0
python-multipart==0.0.6

# Simplified binary analysis libraries (lightweight)
capstone==5.0.0.post1
keystone-engine==0.9.2
requests==2.31.0
psutil==5.9.6
packaging==23.2